+++ hutool-core/src/test/java/com/xiaoleilu/hutool/core/util/TypeUtilTest.java
@@ -1,21 +1,19 @@
 package com.xiaoleilu.hutool.core.util;
 
+import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.junit.Assert;
 import org.junit.Test;
 
 import com.xiaoleilu.hutool.lang.Console;
+import com.xiaoleilu.hutool.util.ReflectUtil;
 import com.xiaoleilu.hutool.util.TypeUtil;
 
 public class TypeUtilTest {
-	
-	@Test
-	public void getTypeTest(){
-		
-	}
 	
 	@Test
 	public void getArgumentTypeTest(){
@@ -27,5 +25,21 @@ public class TypeUtilTest {
 		
 		Type typeArgument = TypeUtil.getTypeArgument(list.getClass());
 		Console.log(typeArgument);
+	}
+	
+	@Test
+	public void getParamTypeTest() {
+		Method method = ReflectUtil.getMethod(TestClass.class, "intTest", Integer.class);
+		Type type = TypeUtil.getParamType(method, 0);
+		Assert.assertEquals(Integer.class, type);
+		
+		Type returnType = TypeUtil.getReturnType(method);
+		Assert.assertEquals(Integer.class, returnType);
+	}
+	
+	public static class TestClass {
+		public Integer intTest(Integer integer) {
+			return 1;
+		}
 	}
 }