+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * 按照Ali规范整理代码。Cache模块中定时清理改为定时任务池方式
 * Http模块全局Header支持
 * JSON模块增加JSONStrFromater，用于格式化非标准的JSON字符串
+* 增加BeanDesc代替JDK的BeanInfo，提供更灵活的Getter和Setter
 
 ### Bug修复
 * db模块修复Oracle中传入java.util.Date对象无法识别类型问题（默认按照Timestamp处理）（感谢@【杭州】wiley）